requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb 
tantivy
streamlit 
pandas
yfinance
openai 
duckduckgo-search 
yfinance 
sqlalchemy 
fastapi[standard]

